Add tests from the base class to the main test classes
Make ListIndexProxyIntegrationTest
and ProofListIndexProxyIntegrationTest inherit from the base class
with tests, applicable to both types of lists, so that no separate
classes are needed.
