--- exonum-java-binding-core/src/test/java/com/exonum/binding/storage/indices/ListIndexIntegrationTest.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2018 The Exonum Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.exonum.binding.storage.indices;
-
-import com.exonum.binding.common.serialization.StandardSerializers;
-import com.exonum.binding.storage.database.View;
-
-/**
- * A test of common ListIndex methods.
- */
-class ListIndexIntegrationTest
-    extends BaseListIndexIntegrationTest {
-
-  @Override
-  ListIndexProxy<String> create(String name, View view) {
-    return ListIndexProxy.newInstance(name, view, StandardSerializers.string());
-  }
-
-  @Override
-  Object getAnyElement(AbstractListIndexProxy<String> index) {
-    return index.get(0L);
-  }
-
-}
